No bug to fix. Read whole comment.
The problem was that you weren't considering CHAR(NR), but VARCHAR2. So
when you were inserting locally you would store the string value without
extra blank spaces (as it was inserted in the database) so when you
immediately tried to delete the inserted row there was no row to match.
(this problem didn't appear -d'oh- when you were deleting a row that you
selected from the database, because locally you would have the same
value as in the database - with the extra blank spaces also -).

Maybe consider the case for CHAR? Or just don't use this in the
database. Seems it only produces problems where you would not expect.